+++ LookingGlass/CommandMenu/ButtonsToCloseMenu.cs
@@ -15,6 +15,7 @@ namespace LookingGlass.EscapeToCloseMenu
         public Rewired.Player player;
         public HGButton closeButton;
         public NetworkUIPromptController networkUIPromptController;
+        public Interactor interactor;
         public void Start()
         {
             //ScrapperPickerPanel(Clone)/MainPanel/Juice/CancelButton
@@ -31,48 +32,53 @@ namespace LookingGlass.EscapeToCloseMenu
             {
                 player = networkUIPromptController.currentLocalParticipant.inputPlayer;
             }
+            if (networkUIPromptController.currentParticipantMaster != null)
+            {
+                player = networkUIPromptController.currentLocalParticipant.inputPlayer;
 
+            }
 
         }
         public void Update()
         {
-            //If not Mouse1
-            //If not Mouse2
-            //If not Interact
-            if (closeButton && !Input.GetMouseButtonDown(0) && !Input.GetMouseButtonDown(1) && !player.GetButton(5) && Input.anyKeyDown)
+            if (player.GetButton(5)) //Interact Key
+            {
+                return;
+            }
+            if (Input.GetMouseButtonDown(0) || Input.GetMouseButtonDown(1)) //Mouse Keys
+            {
+                return;
+            }
+            if (Input.GetKeyDown(KeyCode.Return)) //Requested
+            {
+                return;
+            }
+            if (Input.GetKeyDown(KeyCode.BackQuote) || Input.GetKeyDown(KeyCode.F2)) //Console opening keys
+            {
+                return;
+            }
+            if (Input.anyKeyDown)
             {
                 CloseMenuAfterFrame();
             }
-            // Interact hold blocker to prevent rappidly opening/closing menu because that breaks the menu
-            /*if (interactHoldBlocker && (!LocalUserManager.GetFirstLocalUser().inputPlayer.GetButton(5) || Time.time >= holdBlockerStartTime + 0.5))
-            {//if blocker is active and  player is not holding interact or it has been more than 0.5 seconds
-                interactHoldBlocker = false; //turn off blocker
-
-                if (pickupPickerController != null) // the command menu destroys the pickupPickerController before thus runs, so must check if it is null
-                {
-                    pickupPickerController.enabled = true;
-                    // toggle the pickupPickerController like this and not with PickupPickerController.available because this way isnt networked, and the networking is what was causing issues
-                }
-            }*/
         }
         public void CloseMenuAfterFrame()
         {
-            closeButton.InvokeClick();
-            /*if (pickupPickerController != null) // the command menu destroys the pickupPickerController before thus runs, so must check if it is null
+            if (closeButton)
             {
-                pickupPickerController.enabled = false;
-                // toggle the pickupPickerController like this and not with PickupPickerController.available because this way isnt networked, and the networking is what was causing issues
-            }*/
-
-            /*interactHoldBlocker = true;
-
-            if (pickupPickerController != null) // the command menu destroys the pickupPickerController before thus runs, so must check if it is null
-            {
-                pickupPickerController.enabled = false;
-                // toggle the pickupPickerController like this and not with PickupPickerController.available because this way isnt networked, and the networking is what was causing issues
+                closeButton.InvokeClick();
             }
-            holdBlockerStartTime = Time.time;
-            */
+            if (networkUIPromptController.currentLocalParticipant != null)
+            {
+                if (networkUIPromptController.currentLocalParticipant.cachedBody)
+                {
+                    if (networkUIPromptController.currentLocalParticipant.cachedBody.TryGetComponent<InteractionDriver>(out var interactor))
+                    {
+                        interactor.currentInteractable = null;
+                        interactor.interactableCheckCooldown = 0.1f;
+                    }
+                }
+            }
         }
 
 
@@ -96,12 +102,12 @@ namespace LookingGlass.EscapeToCloseMenu
         }
         public void Setup()
         {
-            turnOffCommandMenu = BasePlugin.instance.Config.Bind<bool>("Command Settings", "Input Disables Command Prompt", true, "Makes any keyboard input other than [Interact] close the command/scrapper/potential etc menu.\n\nDoesn't work with controllers because you wouldn't be able to use the menu.");
+            turnOffCommandMenu = BasePlugin.instance.Config.Bind<bool>("Command Settings", "Input Disables Command Prompt", true, "Makes any keyboard input other than [Interact] close command/ scrapper / potential / etc. menus.\n\nDoesn't work with controllers because you wouldn't be able to use the menu.");
             SetupRiskOfOptions();
         }
         public void SetupRiskOfOptions()
         {
-            ModSettingsManager.AddOption(new CheckBoxOption(turnOffCommandMenu, new CheckBoxConfig() { restartRequired = false }));
+            ModSettingsManager.AddOption(new CheckBoxOption(turnOffCommandMenu, new CheckBoxConfig() { name = "Input closes Command/Pickup menus", restartRequired = false }));
         }
         public void AddCloser(GameObject panelInstance, NetworkUIPromptController net)
         {